- added checks for width|height = 0 - added debug print statements - added missing javadoc

+++ src/main/java/com/longlinkislong/gloop/GLFXStage.java
@@ -51,18 +51,20 @@ import static org.lwjgl.glfw.GLFW.GLFW_KEY_UP;
  */
 public class GLFXStage extends GLObject {
 
+    private static final boolean DEBUG;
+
     static {
+        DEBUG = Boolean.getBoolean("debug") && !System.getProperty("debug.exclude", "").contains("glfxstage");
         PlatformImpl.startup(() -> {
         });
     }
 
     private int width;
     private int height;
-    private final GLMat projection = GLMat4F.ortho(-1f, 1f, -1f, 1f, 0f, 1f).asStaticMat();
+    private final GLMat4F projection = GLMat4F.ortho(-1f, 1f, -1f, 1f, 0f, 1f).asStaticMat();
     private volatile EmbeddedWindow stage;
     private EmbeddedSceneInterface emScene;
     private EmbeddedStageInterface emStage;
-    //private volatile boolean isDirty = true;
     private float scaleFactor = 1f;
     private GLTexture texture;
     private volatile ByteBuffer tBuffer;
@@ -190,7 +192,6 @@ public class GLFXStage extends GLObject {
         public void setCursor(CursorFrame cursorFrame) {
             final CursorType cursorType = cursorFrame.getCursorType();
 
-            System.out.println(cursorFrame.getClass());
             switch (cursorType) {
                 case DEFAULT:
                     updateCursor(GLFXCursor.DEFAULT);
@@ -254,6 +255,16 @@ public class GLFXStage extends GLObject {
     };
 
     /**
+     * Sets the projection matrix.
+     *
+     * @param projection the projection matrix.
+     * @since 15.10.15
+     */
+    public void setProjection(final GLMat4 projection) {
+        this.projection.set(projection.asGLMat4F());
+    }
+
+    /**
      * Retrieves the width of the stage.
      *
      * @return the stage width.
@@ -284,7 +295,11 @@ public class GLFXStage extends GLObject {
     public GLFXStage(final GLThread thread, final int width, final int height) {
         super(thread);
 
-        this.resize(width, height);
+        if (width > 0 && height > 0) {
+            this.resize(width, height);
+        } else {
+            throw new IllegalArgumentException("Width and Height must be at least 1!");
+        }
     }
 
     /**
@@ -297,7 +312,11 @@ public class GLFXStage extends GLObject {
     public GLFXStage(final int width, final int height) {
         super();
 
-        this.resize(width, height);
+        if (width > 0 && height > 0) {
+            this.resize(width, height);
+        } else {
+            throw new IllegalArgumentException("Width and Height must be at least 1!");
+        }
     }
 
     private void setSceneImpl(final Scene scene) {
@@ -363,23 +382,32 @@ public class GLFXStage extends GLObject {
      * @since 15.09.21
      */
     public final void resize(final int newWidth, final int newHeight) {
-        this.width = newWidth;
-        this.height = newHeight;
-
-        if (this.emScene != null) {
-            this.emScene.setSize(width, height);
+        if (DEBUG) {
+            System.out.printf("[GLFXStage] Requested resize: <%d, %d>\n", newWidth, newHeight);
         }
+        if (newWidth > 0 && newHeight > 0) {
+            this.width = newWidth;
+            this.height = newHeight;
 
-        if (this.emStage != null) {
-            this.emStage.setSize(width, height);
+            if (this.emScene != null) {
+                this.emScene.setSize(width, height);
+            }
+
+            if (this.emStage != null) {
+                this.emStage.setSize(width, height);
+            }
+        } else if (DEBUG) {
+            System.err.println("[GLFXStage] Resize rejected; width or height is less than 1.");
         }
 
         this.needsRecreate = true;
     }
 
+    //TODO: determine if this should be public
     public final void scroll(final double deltaX, final double deltaY) {
         // TODO: this doesn't support horizontal scrolling! 
         // there must be a better way
+        // NOTE: who has horizontal mouse wheels?
         GLFXStage.this.emScene.mouseEvent(
                 AbstractEvents.MOUSEEVENT_WHEEL, AbstractEvents.MOUSEEVENT_NONE_BUTTON,
                 leftButton, middleButton, rightButton,
@@ -395,12 +423,16 @@ public class GLFXStage extends GLObject {
         if (this.emScene != null) {
             final int neededSize = this.width * this.height * Integer.BYTES;
 
-            if (this.tBuffer == null || neededSize > this.tBuffer.capacity()) {
-                this.tBuffer = ByteBuffer.allocateDirect(neededSize).order(ByteOrder.nativeOrder());
-            }
+            if (neededSize > 0) {
+                if (this.tBuffer == null || neededSize > this.tBuffer.capacity()) {
+                    this.tBuffer = ByteBuffer.allocateDirect(neededSize).order(ByteOrder.nativeOrder());
+                }
 
-            this.tBuffer.rewind();
-            this.emScene.getPixels(this.tBuffer.asIntBuffer(), this.width, this.height);
+                this.tBuffer.rewind();
+                this.emScene.getPixels(this.tBuffer.asIntBuffer(), this.width, this.height);
+            } else if (DEBUG) {
+                System.err.println("[GLFXStage] Request to read 0 bytes ignored!");
+            }
         }
     }
 
@@ -413,10 +445,10 @@ public class GLFXStage extends GLObject {
         newDrawTask().glRun(this.getThread());
     }
 
-    private void updateCursor(GLFXCursor cursor) {
-
-        System.out.println("Set cursor to " + cursor);
-
+    private void updateCursor(final GLFXCursor cursor) {
+        if (DEBUG) {
+            System.out.println("[GLFXStage] Set cursor to: " + cursor);
+        }
         if (this.window == null || this.window.get() == null) {
             cursor.apply(GLWindow.listActiveWindows().get(0));
         } else {
@@ -434,14 +466,18 @@ public class GLFXStage extends GLObject {
     public GLTask newDrawTask() {
         final GLTask bindTex = GLTask.create(() -> {
             if (this.needsRecreate) {
-                if (this.texture != null) {
-                    this.texture.delete();
+                if (this.width > 0 && this.height > 0) {
+                    if (this.texture != null) {
+                        this.texture.delete();
+                    }
+
+                    this.texture = new GLTexture(this.getThread())
+                            .allocate(1, GLTextureInternalFormat.GL_RGBA8, this.width, this.height);
+
+                    this.needsRecreate = false;
+                } else if (DEBUG) {
+                    System.out.printf("[GLFXStage] Ignored invalid request to resize texture to <%d, %d>\n", this.width, this.height);
                 }
-
-                this.texture = new GLTexture(this.getThread())
-                        .allocate(1, GLTextureInternalFormat.GL_RGBA8, this.width, this.height);
-
-                this.needsRecreate = false;
             }
 
             if (this.needsUpdate) {
@@ -461,7 +497,9 @@ public class GLFXStage extends GLObject {
         return GLTask.join(
                 PROGRAM.get().new UseTask(),
                 PROGRAM.get().new SetUniformITask("fxTexture", 0),
-                PROGRAM.get().new SetUniformMatrixFTask("vProj", this.projection),
+                GLTask.create(() -> {
+                    PROGRAM.get().setUniformMatrixF("vProj", this.projection);
+                }),
                 bindTex,
                 vao.get().new DrawArraysTask(GLDrawMode.GL_TRIANGLE_STRIP, 0, 4));
     }
@@ -534,7 +572,7 @@ public class GLFXStage extends GLObject {
                     break;
                 default:
                     if ((key >= GLFW.GLFW_KEY_A && key <= GLFW.GLFW_KEY_Z) || (key >= GLFW.GLFW_KEY_0 && key <= GLFW.GLFW_KEY_9)) {
-                        keyId = key; // they're all the same -\_0_0_/-                        
+                        keyId = key;
                     }
             }
 
@@ -601,8 +639,8 @@ public class GLFXStage extends GLObject {
     public class KeyCharListener implements GLKeyCharListener {
 
         @Override
-        public void charTypePerformed(GLWindow glw, char c) {
-            int mods = GLFXStage.this.ctrl ? AbstractEvents.MODIFIER_CONTROL : 0;
+        public void charTypePerformed(final GLWindow glw, final char c) {
+            final int mods = GLFXStage.this.ctrl ? AbstractEvents.MODIFIER_CONTROL : 0;
             GLFXStage.this.emScene.keyEvent(AbstractEvents.KEYEVENT_TYPED, com.sun.glass.events.KeyEvent.VK_UNDEFINED, new char[]{c}, mods);
         }
     }
@@ -626,7 +664,7 @@ public class GLFXStage extends GLObject {
     public class MouseButtonListener implements GLMouseButtonListener {
 
         @Override
-        public void mouseButtonActionPerformed(GLWindow glw, int button, GLMouseButtonAction action, Set<GLKeyModifier> set) {
+        public void mouseButtonActionPerformed(final GLWindow glw, final int button, final GLMouseButtonAction action, final Set<GLKeyModifier> set) {
             int buttonId = 0;
 
             if (action == GLMouseButtonAction.PRESSED) {
@@ -685,7 +723,7 @@ public class GLFXStage extends GLObject {
     public class MousePositionListener implements GLMousePositionListener {
 
         @Override
-        public void mousePositionActionPerformed(GLWindow glw, double x, double y) {
+        public void mousePositionActionPerformed(final GLWindow glw, final double x, final double y) {
             if (GLFXStage.this.emScene == null) {
                 return;
             }
@@ -725,25 +763,44 @@ public class GLFXStage extends GLObject {
         }
     }
 
+    /**
+     * StageResizeListener is a GLFramebufferResizeListener that translates
+     * framebuffer resizes to the underlying JavaFX scene. These framebuffer
+     * resizes indicate that the window was resized.
+     *
+     * @since 15.10.15
+     */
     public class StageResizeListener implements GLFramebufferResizeListener {
 
         @Override
-        public void framebufferResizedActionPerformed(GLWindow glw, GLViewport view) {
+        public void framebufferResizedActionPerformed(final GLWindow glw, final GLViewport view) {
             GLFXStage.this.resize(view.width, view.height);
         }
     }
 
+    /**
+     * A MouseScrollListener that translates scrolling to the underlying JavaFX
+     * scene.
+     *
+     * @since 15.10.15
+     */
     public class StageScrollListener implements GLMouseScrollListener {
 
         @Override
-        public void mouseScrollActionPerformed(GLWindow glw, double x, double y) {
+        public void mouseScrollActionPerformed(final GLWindow glw, final double x, final double y) {
             GLFXStage.this.scroll(x, y);
         }
     }
     private Reference<GLWindow> window = null;
     private final Set<Object> activeListeners = new HashSet<>();
 
-    public void addEvents(GLWindow window) {
+    /**
+     * Adds all GLFXStage events to a GLWindow and sets the GLFXStage window.
+     *
+     * @param window the GLWindow
+     * @since 15.10.15
+     */
+    public void addEvents(final GLWindow window) {
         Objects.requireNonNull(window);
         this.window = new WeakReference<>(window);
 
@@ -764,7 +821,14 @@ public class GLFXStage extends GLObject {
         window.addWindowResizeListener(fbListener);
     }
 
-    public void removeEvents(GLWindow window) {
+    /**
+     * Scans the GLWindow for all events derived from this GLFXStage and removes
+     * them.
+     *
+     * @param window the GLWindow to remove the events from.
+     * @since 15.10.15
+     */
+    public void removeEvents(final GLWindow window) {
         if (this.window == null) {
             return;
         } else if (this.window.get() != window) {
